requirements.txt
# Misaka MediaFlow Python Dependencies

# ---- Web 框架 ----
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# ---- 数据库 ----
sqlalchemy>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0                  # PostgreSQL 异步驱动
pymysql>=1.1.0                   # MySQL 同步驱动 (建库用)
aiomysql>=0.2.0                  # MySQL 异步驱动

# ---- HTTP / 缓存 ----
httpx>=0.27.0
redis>=5.0.0

# ---- 调度 / 配置 ----
apscheduler>=3.10.0
pyyaml>=6.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

